4-records-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23

pydantic==2.5.0
requests==2.31.0
strawberry-graphql>=0.216.0